Ya hay foto de la aplicación + muchos arreglos pequeños

+++ app/src/main/java/com/example/homehub/Expenses.java
@@ -2,7 +2,7 @@ package com.example.homehub;
 
 public class Expenses {
     private String name;
-    private int quantity;
+    private double quantity;
 
     public String getName() {
         return name;
@@ -11,15 +11,16 @@ public class Expenses {
     public void setName(String name) {
         this.name = name;
     }
-    public int getQuantity() {
+
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
 
-    public Expenses(String name, int quantity){
+    public Expenses(String name, double quantity) {
         this.name = name;
         this.quantity = quantity;
     }
